filter_rewrite: Migrated RegisterBlueprintsMixin, StartupMixin and UpdatesMixin into Teleflask.
